+++ lib/src/widgets/waveform_display.dart
@@ -1,9 +1,10 @@
 import 'package:audio_waveform_kit/src/controllers/audio_recording_bloc.dart';
 import 'package:audio_waveform_kit/src/painters/waveform_painter.dart';
+import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-enum WaveformStyle { envelope, string }
+const _listEquality = ListEquality<double>();
 
 class WaveformDisplay extends StatelessWidget {
   const WaveformDisplay({
@@ -59,7 +60,7 @@ class WaveformDisplay extends StatelessWidget {
   ) {
     if (prev is AudioRecordingState$Recording &&
         curr is AudioRecordingState$Recording) {
-      return prev.waveformSamples != curr.waveformSamples;
+      return !_listEquality.equals(prev.waveformSamples, curr.waveformSamples);
     }
     return prev.runtimeType != curr.runtimeType;
   }